# backend/entityservice/api_def/openapi.yaml
...
openapi: 3.0.0
info:
  version: '1.13'
  title: Entity Matching API
  x-logo:
    url: http://clkhash.readthedocs.io/en/latest/_static/logo.svg
  contact:
    name: 'Confidential Computing, Data61 | CSIRO'
    email: confidential-computing@csiro.au
    url: https://github.com/data61/anonlink-entity-service

  description: >-
    Allows multiple organisations to carry out private record linkage -
    without disclosing personally identifiable information.

    In general:

    * All parameters and returned objects are JSON with content-type set to
    `"application/json"`.

    * All authentication tokens are 48 character hex strings.

    * The `Project`/`Run` resource identifiers are also 48 char hex strings.

    * HTTP status codes are used to distinguish server outcomes.


    **Note:** Several important concepts including security considerations are described in
    [separate documentation](./concepts.html).

    ### Introduction

    This is a service that computes matches of entity records between two parties.
    The personally identifiable information used for linking is first locally transformed
    to anonymous linking codes called [Cryptographic Longterm Keys](concepts.html#cryptographic-longterm-key)
    (CLKs). Each party then uploads its CLKs to the service for matching.
    The service supports three different [types of outputs](concepts.html#result-types)
    of matching results, varying in privacy properties.
...
    and uploading the CLKs.


    ### Security

...

...
servers:
...
paths:
...
  /projects:
    get:
      operationId: entityservice.views.project.projects_get
      summary: List of record linkage projects.
      tags:
        - Project
      description: |
        List of all privacy preserving record linkage projects.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectList'
    post:
      operationId: entityservice.views.project.projects_post
      summary: Create and configure a new linkage project
      tags:
        - Project
      description: |
        Prepare resources and endpoints to accept CLK uploads from data providers.
        Provides the caller with credentials that will be required for any further
        interaction with this project.

        ### Schema

        Although these comprise the column names, the raw data will never be
        sent to this entity service.

        Each participant will be able to see the schema to verify it is
        what they expect. Schema details should have been determined and agreed
        on by each party before starting a mapping task. This is documented
        in [schema](./concepts.html#schema).

        ### Result Type

        The result type specifies what information is available after the entity
        resolving process has completed. All project **runs** will use this result type.

        The parts of the computed linkage results that are accessable by the different tokens depends on the
        `result_type`:

        - `"similarity_scores"`, or `"groups"`\
        If the `result_type` is `"similarity_scores"` or `"groups"` then the results can be accessed with the
        `result_token``token`, which is provided when initially creating the mapping.

        - `"permutations"`\
        If the `result_type` is `permutations`, then the data providers can access their respective permutation with
        their individual `receipt_token`, which they obtain when adding data to the mapping.
        The mask can be accessed with the `result_token`.

        Only `"groups"` supports multi-party linkage. If the result type is `"similarity_scores"` or
        `"permutations"`, then the number of parties must be 2.

      requestBody:
        description: new project info
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewProject'
      responses:
        '201':
          description: New project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewProjectResponse'
        '400':
          $ref: "#/components/responses/BadRequest"
        '403':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/RateLimited'
  '/projects/{project_id}':
    parameters:
      - $ref: '#/components/parameters/token'
      - $ref: '#/components/parameters/project_id'

    get:
      operationId: entityservice.views.project.project_get
      summary: Description of a particular linkage project
      tags:
        - Project
      description: |
        Details provided at project creation.

      responses:
        '200':
          description: Project description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDescription'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/RateLimited'

    delete:
      operationId: entityservice.views.project.project_delete
      summary: Delete linkage project.
      tags:
        - Project
      description: Remove the project and all associated data from the service.
      responses:
        '204':
          description: Successfully deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/RateLimited'
...
  '/projects/{project_id}/clks':
    post:
      operationId: entityservice.views.project.project_clks_post
      summary: Upload encoded PII data to a linkage project.
      tags:
        - Project
      description: |
        Called by each of the data providers with their calculated `CLK` vectors.
        The project must have been created, and the caller must have both the
        `project_id` and a valid `upload_token` in order to contribute data.

        The data uploaded must be of one of the following formats.
        - CLKs only upload: An array of base64 encoded [CLKs](./concepts.html#cryptographic-longterm-keys), one per
          entity.
        - CLKs with blocking information upload: An array of base64 encoded CLKs with corresponding blocking
          information. One element in this array is an array with the first element being a base64 encoded CLK followed
          by the block IDs the CLK corresponds to.

        When the last party successfully uploads their data any existing project runs are queued
        for execution.

        The uploaded encodings must all have the same length in bytes. If the project's linkage schema
        specifes an encoding size it will be checked and enforced before any runs are computed. Note a
        minimum and maximum encoding size can be set at the server level at deployment time.
        Currently anonlink requires this _encoding size_ to be a multiple of 8. An example value is 128 Bytes.

        Note in the default deployment the maximum request size is set to `~10 GB`, which __should__
        translate to just over 20 million entities.

        ### JSON Upload

        This endpoint can be used with the Content-Type: application/json and uses the `CLKUpload`
        structure of a JSON array of base64 encoded strings.

        ### Binary Upload

        An additional api endpoint (/projects/{project_id}/binaryclks) has been added for uploading CLKs as a binary
        file. This is to allow for faster and more efficient data transfer.

      parameters:
        - $ref: '#/components/parameters/project_id'
        - $ref: '#/components/parameters/token'
      requestBody:
        description: the encoded PII
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/CLKUpload'
                - $ref: '#/components/schemas/CLKnBlockUpload'
      responses:
        '201':
          description: Data Uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadReceipt'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/RateLimited'
  '/projects/{project_id}/binaryclks':
    post:
      operationId: entityservice.views.project.project_binaryclks_post
      summary: Upload binary encoded PII data to a linkage project.
      tags:
        - Project
      description: |
        An experimental api for uploading CLKs as a binary file. This is to allow for
        faster and more efficient data transfer.
        Called by each of the data providers with their calculated `CLK` vectors.
        The project must have been created, and the caller must have both the
        `project_id` and a valid `upload_token` in order to contribute data.

        When the last party successfully uploads their data any existing project runs are queued
        for execution.

        The uploaded encodings must all have the same length in bytes. If the project's linkage schema
        specifes an encoding size it will be checked and enforced before any runs are computed. Note a
        minimum and maximum encoding size can be set at the server level at deployment time.
        Currently anonlink requires this _encoding size_ to be a multiple of 8. An example value is 128 Bytes.

        Additional headers are required:

          Content-Type: application/octet-stream
          Hash-Count: <number of CLKS>
          Hash-Size: <size of one CLK in bytes>

        The file is a series of CLK data:

            bit_packing_fmt = "!<Hash-Size>s"

        Where:

            "!" Use network byte order (big-endian).
            "<Hash-Size>s" Store the 128 raw bytes of the bitarray. Default value is "128s".

        https://docs.python.org/3/library/struct.html#format-strings

      parameters:
        - $ref: '#/components/parameters/project_id'
        - $ref: '#/components/parameters/token'
        - in: header
          name: Hash-Count
          required: true
          schema:
            type: integer
        - in: header
          name: Hash-Size
          required: true
          schema:
            type: integer
      requestBody:
        description: the clks in binary
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary

      responses:
        '201':
          description: Data Uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadReceipt'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/RateLimited'
...
components:
  parameters:
    token:
      required: true
      schema:
        type: string
      name: Authorization
      in: header
    project_id:
      in: path
      name: "project_id"
      description: Resource identifier of the record linkage project
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: An unspecified error occured
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    BadRequest:
      description: Client's request could not be processed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Authentication token could be missing, or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    RateLimited:
      description: Client is rate limited. Try requesting less frequently.
  schemas:
...
    ProjectList:
      type: array
      items:
        type: object
        description: Listing of projects.
        properties:
          project_id:
            type: string
          time_added:
            type: string
            description: ISO8601 formated datetime
            format: 'date-time'
    Schema:
      type: object
      description: |
        The `schema` is described in the [clkhash](http://clkhash.readthedocs.io/en/latest/schema.html)
        documentation.

        A json-schema definition can be found [here](http://clkhash.readthedocs.io/en/latest/_static/schema.json#hashing-schema)
    ResultType:
      type: string
      description: |
        Defines the output type of the mapping. Multi-party linkage requires `"groups"` to be used.
      enum:
        - groups
        - permutations
        - similarity_scores
...
    NewProject:
      description: Configuration for a new linkage project.
      type: object
      x-body-name: project
      properties:
        schema:
          $ref: '#/components/schemas/Schema'
        result_type:
          $ref: '#/components/schemas/ResultType'
        number_parties:
          description: |
            How many data providers will participate in this project. Default value is 2.
            If there are more than two data providers, then `result_type` must be
            `"groups"`.
          type: integer
          minimum: 2
        name:
          description: Any free text name for this project.
          type: string
        notes:
          description: Any free text to store with this project.
          type: string
      required:
        - schema
        - result_type
    ProjectDescription:
      allOf:
        - $ref: '#/components/schemas/NewProject'
        - type: object
          properties:
            project_id:
              type: string
            parties_contributed:
              type: number
            error:
              type: boolean
    NewProjectResponse:
      properties:
        project_id:
          type: string
        update_tokens:
          type: array
          description: Array of unique tokens authorizing the upload of CLKs.
          items:
            type: string
        result_token:
          description: A token required to access mapping results.
          type: string
...
    CLKUpload:
      description: Object that contains this party's Bloom Filters
      type: object
      required: [clks]
      properties:
        clks:
          type: array
          items:
            type: string
            format: byte
            description: Base64 encoded CLK data
    CLKnBlockUpload:
      description: Object that contains this party's Bloom Filters including blocking information
      type: object
      required: [clknblocks]
      properties:
        clknblocks:
          type: array
          items:
            type: array
            items:
              anyOf:
                - type: string
                  format: byte
                  description: Base64 encoded CLK data
                - type: string
                - type: integer
    UploadReceipt:
      properties:
        receipt_token:
          type: string
        message:
          type: string
...